# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "图书",
        "version": "1"
    },
    "paths": {
        "/api/book/get/{id}": {
            "get": {
                "operationId": "App\\Http\\Controllers\\BookController::getBookById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回根据id查询的书"
                    }
                }
            }
        },
        "/api/allBook": {
            "get": {
                "operationId": "App\\Http\\Controllers\\BookController::getAllBook",
                "responses": {
                    "200": {
                        "description": "查询所有图书"
                    }
                }
            }
        },
        "/api/book/price/{sort}": {
            "get": {
                "operationId": "App\\Http\\Controllers\\BookController::getBookByPrice",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询图书,传参为1从贵到便宜，传参其他从便宜到贵"
                    }
                }
            }
        },
        "/api/books/price/{sort}": {
            "get": {
                "operationId": "App\\Http\\Controllers\\BookController::getBooksByPrice",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询图书,传参为1从贵到便宜，传参其他从便宜到贵"
                    }
                }
            }
        },
        "/api/swagger/doc": {
            "get": {
                "operationId": "App\\Http\\Controllers\\SwaggerController::doc",
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}